requirements.txt
apache_beam[dataframe]
apache-beam[gcp]
google-cloud-storage
google-cloud-bigquery